Wolfram|Alpha: fix recalculate responses
The recalculate response doesn't exclude the Input pod. Fix our pod
defaults to skip Input if we can. Also fix the affinity for Results, as
the pod is named Result.

We also need to stop excluding Input and selecting pod index 1, because
the recalculate URL preserves the podindex but doesn't preserve the
exclusion, meaning we end up just getting our input interpretation back.
Instead, always ask for pods 1 and 2, and just filter ourselves.